fix #145 add REPL links to landing page examples
