+++ src/components/Examples.svelte
@@ -47,6 +47,7 @@
     <LanguageSlot let:option {option} slot="option" />
   </MultiSelect>
   <CollapsibleCode
+    repl_url="https://svelte.dev/repl/e3b88f59f62b498d943ecf7756ab75d7"
     code={`
 <pre>bind:selected = {JSON.stringify($language_store)}</pre>
 
@@ -84,6 +85,7 @@
     {loading}
   />
   <CollapsibleCode
+    repl_url="https://svelte.dev/repl/79e22e1905c94456aa21564b4d5f8759"
     code={`
 <pre>selected = {JSON.stringify(selected_ml)}</pre>
 
@@ -128,6 +130,7 @@
   {/if}
 
   <CollapsibleCode
+    repl_url="https://svelte.dev/repl/516279bd62ec424986115263c2cdc169"
     code={`
 <label for="confetti-select">Callback on item selection</label>
 
@@ -185,6 +188,7 @@
     </p>
   </form>
   <CollapsibleCode
+    repl_url="https://svelte.dev/repl/3a217c39932047a09f61d6425b04a7c3"
     code={`
 <label for="color-select">
   Color select using the \`selected\` and \`option\` slot components to render colors.
@@ -230,7 +234,7 @@
 </section> -->
 <style>
   section {
-    margin-top: 2em;
+    margin-bottom: 2em;
     background-color: #28154b;
     border-radius: 4pt;
     padding: 1pt 10pt;
